+++ src/inspectors/call_graph/mod.rs
@@ -318,4 +318,17 @@ mod tests {
             )
         );
     }
+
+    #[test]
+    fn call_graph_for_public_getter() {
+        let inspector = CallGraphInspector::new(fixture_project());
+        let id = make_id("Getter", "project");
+        let output = inspector.inspect(&id).unwrap().to_string();
+        assert_eq!(
+            output,
+            include_str!(
+                "../../../fixtures/inspect-call-graph/expected/call_graph_for_public_getter.txt"
+            )
+        );
+    }
 }